Fix User entity and config

+++ src/main/java/it/intesys/codylab/db/model/User.java
@@ -9,7 +9,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String name;
     private String surname;
@@ -29,11 +29,11 @@ public class User {
     public User() {
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public User setId(Long id) {
+    public User setId(Integer id) {
         this.id = id;
         return this;
     }
